bionic: remove bzero() fall-through to memset()
The original ENTRY/END semantics can cause the fall-through in (or likely
the return from) bzero() to fail when using the Cortex-optimized assembly
code from Linaro. For now, just reimplement bzero() as memset() with r1
(character to set) fixed to "0".

(Possible FIXME: properly call memset(), but this is likely faster
as it's really just inlining memset() anyway)

Change-Id: If04841c705d2beed58a57a05ecfd254e3f595562